# perf: Improve `approx_topk` performance by reducing allocations.