--- a/pkg/logql/count_min_sketch.go
+++ b/pkg/logql/count_min_sketch.go
@@ -201,7 +201,7 @@ func (v *HeapCountMinSketchVector) Add(metric labels.Labels, value float64) {
 		heap.Push(v, metric)
 		v.observed[id] = struct{}{}
 	} else if labels.Equal(v.Metrics[0], metric) {
-		// The smalles element has been updated to fix the heap.
+		// The smallest element has been updated to fix the heap.
 		heap.Fix(v, 0)
 	}