… vendor/github.com/docker/go-metrics/namespace.go