@cstyan 2024-12-17
So there's no actual perf. change here now. The benchmark is worth including, but how are we getting to a state where the  len/cap of the slice is going beyond the max we set in the first place?